Remove shared behaviour for custom hooks in order for it to leave on resource custom hooks file
