--- internal/serviceapi/privatelinkendpointservicedatafederationonlinearchiveapi/hooks_shared.go
@@ -1,13 +1,0 @@
-package privatelinkendpointservicedatafederationonlinearchiveapi
-
-func normalizeOptionalStringFields(obj map[string]any) {
-	setEmptyStringIfMissing(obj, "comment")
-	setEmptyStringIfMissing(obj, "region")
-	setEmptyStringIfMissing(obj, "customerEndpointDNSName")
-}
-
-func setEmptyStringIfMissing(obj map[string]any, responseKey string) {
-	if val, exists := obj[responseKey]; !exists || val == nil {
-		obj[responseKey] = ""
-	}
-}
