+++ internal/serviceapi/privatelinkendpointservicedatafederationonlinearchiveapi/resource_custom_hooks.go
@@ -151,3 +151,15 @@ func prepareBody(bodyReq []byte) ([]byte, bool) {
 	}
 	return modifiedBody, true
 }
+
+func normalizeOptionalStringFields(obj map[string]any) {
+	setEmptyStringIfMissing(obj, "comment")
+	setEmptyStringIfMissing(obj, "region")
+	setEmptyStringIfMissing(obj, "customerEndpointDNSName")
+}
+
+func setEmptyStringIfMissing(obj map[string]any, responseKey string) {
+	if val, exists := obj[responseKey]; !exists || val == nil {
+		obj[responseKey] = ""
+	}
+}
